requirements.txt
# --- Core AI & LangChain ---
langchain>=0.3.0
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-chroma
langchain-text-splitters

# --- Vector Database & Embeddings ---
chromadb
sentence-transformers
huggingface-hub

# --- API & Web Interface ---
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
streamlit
requests

# --- Evaluation (RAGAS) ---
ragas>=0.2.1
datasets
pandas
nest-asyncio

# --- Utilities ---
python-dotenv
PyPDF2
tabulate